# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TTS Storyteller API",
    "description": "Multi-backend TTS API supporting Qwen3-TTS and VibeVoice for voice-driven story generation with parallel execution",
    "version": "0.1.0"
  },
  "paths": {
...
    "/voices": {
      "get": {
        "summary": "List Voices",
        "description": "List all available voices, optionally filtered by pool.\n\nArgs:\n    pool: Optional pool name to filter voices",
        "operationId": "list_voices_voices_get",
        "parameters": [
          {
            "name": "pool",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pool"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Voice"
                  },
                  "title": "Response List Voices Voices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Voice Endpoint",
        "description": "Create a new voice using voice design (Qwen only).\n\nGenerates voice from text description. Not supported for VibeVoice backend.\nFor VibeVoice, use POST /voices/clone instead.\n\nGenerates WAV and prompt files and saves the voice to the database.\nReturns a job ID to track generation progress.",
        "operationId": "create_voice_endpoint_voices_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceConfig"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/voices/{voiceId}": {
      "get": {
        "summary": "Get Voice Endpoint",
        "description": "Get voice details by ID.",
        "operationId": "get_voice_endpoint_voices__voiceId__get",
        "parameters": [
          {
            "name": "voiceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,100}$",
              "title": "Voiceid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voice"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Voice Endpoint",
        "description": "Update an existing voice configuration.\n\nIf config changed, triggers regeneration job.\nOtherwise, just saves the updated configuration.",
        "operationId": "update_voice_endpoint_voices__voiceId__put",
        "parameters": [
          {
            "name": "voiceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,100}$",
              "title": "Voiceid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Voice Endpoint",
        "description": "Delete a voice.\n\nRemoves the voice from the database and all pools.\nDoes not delete WAV or prompt files.",
        "operationId": "delete_voice_endpoint_voices__voiceId__delete",
        "parameters": [
          {
            "name": "voiceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,100}$",
              "title": "Voiceid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/voices/clone": {
      "post": {
        "summary": "Clone Voice Endpoint",
        "description": "Create a new voice using voice cloning from reference audio.\n\nWorks with all backends (qwen, vibevoice). Requires reference audio file.\nGenerates a prompt from the reference audio and saves the voice to the database.\nReturns a job ID to track generation progress.",
        "operationId": "clone_voice_endpoint_voices_clone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceCloneConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Voice": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,100}$",
            "title": "Id"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Language for this voice"
          },
          "instruction": {
            "type": "string",
            "title": "Instruction",
            "description": "Voice instruction/prompt"
          },
          "sample_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Text",
            "description": "Sample text for this voice"
          },
          "backend": {
            "type": "string",
            "title": "Backend",
            "description": "TTS backend (qwen, vibevoice)",
            "default": "qwen"
          },
          "promptPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promptpath",
            "description": "Path to prompt file"
          },
          "refAudioPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refaudiopath",
            "description": "Path to reference audio"
          }
        },
        "type": "object",
        "required": [
          "id",
          "language",
          "instruction"
        ],
        "title": "Voice",
        "description": "A voice definition."
      },
      "VoiceCloneConfig": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,100}$",
            "title": "Id",
            "description": "Voice identifier"
          },
          "language": {
            "type": "string",
            "minLength": 1,
            "title": "Language",
            "description": "Language for this voice"
          },
          "instruction": {
            "type": "string",
            "minLength": 1,
            "title": "Instruction",
            "description": "Voice description/notes"
          },
          "ref_audio_url": {
            "type": "string",
            "minLength": 1,
            "title": "Ref Audio Url",
            "description": "Path to reference audio file (WAV format). Upload via POST /audio/upload first to get the path."
          },
          "ref_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Text",
            "description": "Optional transcript of reference audio"
          },
          "backend": {
            "type": "string",
            "title": "Backend",
            "description": "TTS backend to use (qwen, vibevoice)",
            "default": "qwen"
          }
        },
        "type": "object",
        "required": [
          "id",
          "language",
          "instruction",
          "ref_audio_url"
        ],
        "title": "VoiceCloneConfig",
        "description": "Voice configuration for voice cloning from reference audio (works with all backends).\n\nWorkflow:\n1. Upload reference audio: POST /audio/upload (returns file_path)\n2. Use file_path in ref_audio_url field\n3. Create voice clone: POST /voices/clone"
      },
      "VoiceConfig": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,100}$",
            "title": "Id",
            "description": "Voice identifier"
          },
          "language": {
            "type": "string",
            "minLength": 1,
            "title": "Language",
            "description": "Language for this voice"
          },
          "instruction": {
            "type": "string",
            "minLength": 1,
            "title": "Instruction",
            "description": "Voice instruction/prompt"
          },
          "sample_text": {
            "type": "string",
            "minLength": 1,
            "title": "Sample Text",
            "description": "Sample text for voice generation"
          },
          "backend": {
            "type": "string",
            "title": "Backend",
            "description": "TTS backend to use (qwen only - vibevoice not supported for voice design)",
            "default": "qwen"
          }
        },
        "type": "object",
        "required": [
          "id",
          "language",
          "instruction",
          "sample_text"
        ],
        "title": "VoiceConfig",
        "description": "Voice configuration for creation/update using voice design (Qwen only)."
      },
...
    }
  }
}